requirements.txt
jieba==0.42.1
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
scikit-learn==1.1.3
torch==1.10.1
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.11.2